Base the throttling on careful(ish) study of the ircd...
The first four lines can be sent as fast as you like... after that, there
must be at least two seconds (ala time()!) before you send the next or a
two-second delay will be inserted for you.

Try to hack this, but the inter-line delay appears to need to be three
seconds because of the granularity of time().

Maybe the ircd should switch to a better time source...