New script (e.g. to be used with JSexec) for the purpose of automated posting to message sub-boards - a preferred alternative to using "smbutil i". The script closely mimics smbutil with an important distinction: To import a file for the message body text, use the '-i<filename>' option. Otherwise, stdin will be read for the message body. Like smbutil, if the -t, -f, -s, or -d options aren't specified, one or more required header values (to, from, subject) may be prompted for.
The main advantage here is that JSexec knows enough detail about the system
configuration to generate valid/unique Message-IDs for the newly created
messages (unlike smbutil). Message-IDs are critical for message tracing and
dupe-detection on networked message subs. Please use this script to post
(in lieu of smbutil) automated messages to networked subs.