Have TickIT use the "closest" address in the From line.
This may resolve Dumas Walker's issue with cross-domain TIC forwarding.
This should really use the value from BinkpSourceAddress, but that would
mean renaming that key and including it in the TicITCfg object, or
creating a new SBBSEchoCfg object method or something, and I don't want
to mess with the sbbsecho.ini format.