This is a script that can sort a file.  This script is meant to be run with jsexec (on the OS command prompt, not via the Synchronet terminal server).  One purpose of this script is to sort dictionary files when they're created or when words are added to existing dictionary files.
The command-line syntax of this script is:
sexec sort_file.js <inputFilename> <outputFilename> [convertToLowercase]
Where <inputFilename> is the name of the input file and <outputFilename> is the name of the output file (the sorted input).  convertToLowercase is optional, and specifies whether or not to convert all entries in the file to lower-case (which is required for case-insensitive word matching via SlyEdit, for instance).  The values for convertToLowercase can be true or false (really, anything besides true will be interpreted as false).