New modopts.ini option: expiration_after_validation (default: false)
If this value is false (the default), then a validated user's account will
be set to *not* expire.
If this value is true, then a validated user's account expiration status
won't be changed unless the expiration_days_after_validation option is also
set, in which case the expiration date will be reset based on that setting.

Document the new modopts.ini [logon] email_validation option to enable use of
this module.